Compose: Allow modules to remove compose sequences
With this patch, the content of the .remove-files corresponding to
the selected modules are deleted from the compiled Compose file. This
allows for a much greater control when overwriting compose sequences
without relying on certain behaviour of the parsing applications (Does
the first or the last occurence count?). Moreover, we can now delete
prefixes of sequences we want to use.

The resulting Compose will be whitespace-normalized (whitespace outside
of quoted words will be replaced by a single space), stripped of keysyms
and comments, sorted and deduplicated.

Signed-off-by: hrnz <ulli@hrnz.li>