Eliminate the need for perl.
Replace the only perl script in "grafik" by a python script.
Unfortunately this requires the Python package "lxml" (which is
commonly used, though), since the "xml" package from the Python
standard library is not quite capable.

I put special care in the Python script to generate the same svg as
the perl script - although both are not identical. Differences are
e.g. trailing zeros in fractional numbers and slightly different xml
formatting.